website: add back `limactl sudoers`
Add back https://lima-vm.io/docs/reference/limactl_sudoers/

Also clarified that it is specific to macOS, and fixed the doc URL

Fix issue 3436
