+++ cmd/limactl/start-at-login_unix.go
@@ -16,24 +16,6 @@ import (
 	"github.com/spf13/cobra"
 )
 
-func startAtLoginCommand() *cobra.Command {
-	startAtLoginCommand := &cobra.Command{
-		Use:               "start-at-login INSTANCE",
-		Short:             "Register/Unregister an autostart file for the instance",
-		Args:              WrapArgsError(cobra.MaximumNArgs(1)),
-		RunE:              startAtLoginAction,
-		ValidArgsFunction: startAtLoginComplete,
-		GroupID:           advancedCommand,
-	}
-
-	startAtLoginCommand.Flags().Bool(
-		"enabled", true,
-		"Automatically start the instance when the user logs in",
-	)
-
-	return startAtLoginCommand
-}
-
 func startAtLoginAction(cmd *cobra.Command, args []string) error {
 	instName := DefaultInstanceName
 	if len(args) > 0 {
@@ -70,8 +52,4 @@ func startAtLoginAction(cmd *cobra.Command, args []string) error {
 	}
 
 	return nil
-}
-
-func startAtLoginComplete(cmd *cobra.Command, _ []string, _ string) ([]string, cobra.ShellCompDirective) {
-	return bashCompleteInstanceNames(cmd)
 }
